refactor: Button 컴포넌트 가독성 개선, ui 폴더에 constants.js 생성
