+++ src/components/layout/Header.jsx
@@ -14,7 +14,6 @@ export default function Header() {
         </Link>
         {isHome && (
           <Link to='/post' aria-label='만들기 화면으로 이동'>
-            {/*TODO: Button--outlined-40 으로 교체*/}
             <button className={styles['button-outlined']}>
               롤링 페이퍼 만들기
             </button>
