+++ src/components/layout/Header.module.scss
@@ -43,7 +43,6 @@
   padding: rem(8px) $size-md;
   color: $gray900;
   font-size: $font-md;
-  font-weight: bold;
 
   &:disabled {
     background-color: $gray300;
