+++ src/components/ui/Button.jsx
@@ -1,33 +1,32 @@
 import classNames from 'classnames/bind';
 
 import styles from '@/components/ui/Button.module.scss';
+import { BUTTON_SIZE, BUTTON_VARIANT } from '@/components/ui/constants';
 
 export default function Button({
   variant = BUTTON_VARIANT.primary,
   size = BUTTON_SIZE.big,
-  text = '',
   onClick = () => {},
+  children,
   ...rest
 }) {
   const cn = classNames.bind(styles);
   return (
     <button
       onClick={onClick}
-      className={cn(`button-${size}`, {
-        'button-primary': variant === BUTTON_VARIANT.primary,
-        'button-secondary': variant === BUTTON_VARIANT.secondary,
-      })}
+      className={cn(
+        {
+          'button-big': size === BUTTON_SIZE.big,
+          'button-small': size === BUTTON_SIZE.small,
+        },
+        {
+          'button-primary': variant === BUTTON_VARIANT.primary,
+          'button-secondary': variant === BUTTON_VARIANT.secondary,
+        }
+      )}
       {...rest}
     >
-      {text}
+      {children}
     </button>
   );
 }
-export const BUTTON_VARIANT = {
-  primary: 'primary',
-  secondary: 'secondary',
-};
-export const BUTTON_SIZE = {
-  small: 'small',
-  big: 'big',
-};
