+++ src/components/ui/ProfileIcon.jsx
@@ -1,6 +1,7 @@
 import classNames from 'classnames/bind';
 
 import DefaultProfile from '@/assets/images/default_profile.png';
+import { PROFILE_ICON_SIZE } from '@/components/ui/constants';
 import styles from '@/components/ui/ProfileIcon.module.scss';
 
 export default function ProfileIcon({
@@ -10,11 +11,12 @@ export default function ProfileIcon({
   ...rest
 }) {
   const cn = classNames.bind(styles);
+  const isEmptySrc = src === '';
   const onErrorImg = (e) => {
     e.target.src = DefaultProfile;
   };
 
-  return src === '' ? (
+  return isEmptySrc ? (
     <div className={cn('profile', `profile-${size}`, { stroke })} {...rest}>
       <img src={DefaultProfile} alt='기본 프로필 이미지' />
     </div>
@@ -29,7 +31,3 @@ export default function ProfileIcon({
     </div>
   );
 }
-export const PROFILE_ICON_SIZE = {
-  small: 'small',
-  big: 'big',
-};
