+++ src/components/ui/constants.js
@@ -1,0 +1,12 @@
+export const BUTTON_VARIANT = {
+  primary: 'primary',
+  secondary: 'secondary',
+};
+export const BUTTON_SIZE = {
+  small: 'small',
+  big: 'big',
+};
+export const PROFILE_ICON_SIZE = {
+  small: 'small',
+  big: 'big',
+};
